+++ mis-vpay/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Login from './components/Auth/Login';
 import Home from './components/Dashboard/Home';
 import { BrowserRouter, Routes,Route } from 'react-router-dom';
-import Retail_Treansation from "../src/components/Retail/Retail_Treansation"
+import Retail_Transaction from "../src/components/Retail/Retail_Transaction"
 import Table from "../src/components/Table/SalesTable"
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
       <Routes>
         <Route path='/' element={<Login/>}/>
         <Route path='/Home' element={<Home/>}/>
-        <Route path='/Transcation' element={<Retail_Treansation/>}/>
+        <Route path='/Transaction' element={<Retail_Transaction/>}/>
         <Route path='/Table' element={<Table/>}/>
       </Routes>
     </BrowserRouter>
